remove useless plugin
