+++ index.js
@@ -8,7 +8,6 @@ const remark2rehype = require('remark-rehype')
 const inspect = require('unist-util-inspect')
 
 const abbr = require('./packages/abbr')
-const replaceAbbr = require('./packages/replace-abbr')
 const align = require('./packages/align')
 const trailingSpaceHeading = require('./packages/trailing-space-heading')
 const headingShifter = require('./packages/heading-shift')
@@ -38,7 +37,6 @@ const processor = (config) =>
     .use(customBlocks, config.customBlocks)
     .use(align)
     .use(abbr)
-    .use(replaceAbbr)
     .use(math, config.math)
     .use(katex, config.katex)
     .use(htmlBlocks)
