+++ packages/abbr/README.md
@@ -1,1 +1,15 @@
-This plug-in detects abbreviations (`*[abbreviation]: description`) and creates a node abbr.
+This plug-in replaces abbreviations.
+
+```
+This is an abbreviation: REF.
+ref and REFERENCE should be ignored.
+
+*[REF]: Reference
+```
+
+->
+
+```HTML
+<p>This is an abbreviation: "<abbr title="Reference">REF</abbr>".
+ref and REFERENCE should be ignored.</p>
+```
