--- packages/replace-abbr/README.md
@@ -1,15 +1,0 @@
-This plug-in replaces words using abbr nodes.
-
-```
-This is an abbreviation: REF.
-ref and REFERENCE should be ignored.
-
-*[REF]: Reference
-```
-
-->
-
-```HTML
-<p>This is an abbreviation: "<abbr title="Reference">REF</abbr>".
-ref and REFERENCE should be ignored.</p>
-```
